# perf: Increase default cross-file parallelism limit for new-streaming multiscan