--- a/crates/polars-stream/src/nodes/io_sources/multi_file_reader/initialization/slice.rs
+++ b/crates/polars-stream/src/nodes/io_sources/multi_file_reader/initialization/slice.rs
@@ -115,7 +115,7 @@ async fn resolve_negative_slice(config: &MultiFileReaderConfig) -> PolarsResult<
                 PolarsResult::Ok(reader)
             }))
         })
-        .buffered(config.n_readers_pre_init.min(config.sources.len()));
+        .buffered(config.n_readers_pre_init().min(config.sources.len()));
 
     let n_rows_needed = IdxSize::try_from(offset_from_end).unwrap();
     let slice_len_idxsize = IdxSize::try_from(slice_len).unwrap_or(IdxSize::MAX);