--- a/crates/polars-stream/src/nodes/io_sources/multi_file_reader/reader_pipelines/generic.rs
+++ b/crates/polars-stream/src/nodes/io_sources/multi_file_reader/reader_pipelines/generic.rs
@@ -24,11 +24,9 @@ use crate::nodes::io_sources::multi_file_reader::extra_ops::missing_columns::ini
 use crate::nodes::io_sources::multi_file_reader::extra_ops::{
     ExtraOperations, apply_extra_columns_policy,
 };
+use crate::nodes::io_sources::multi_file_reader::initialization::MultiScanTaskInitializer;
 use crate::nodes::io_sources::multi_file_reader::initialization::slice::{
     ResolvedSliceInfo, resolve_to_positive_slice,
-};
-use crate::nodes::io_sources::multi_file_reader::initialization::{
-    MultiScanTaskInitializer, max_concurrent_scans_config,
 };
 use crate::nodes::io_sources::multi_file_reader::post_apply_pipeline::PostApplyPool;
 use crate::nodes::io_sources::multi_file_reader::reader_interface::capabilities::ReaderCapabilities;
@@ -206,7 +204,7 @@ impl MultiScanTaskInitializer {
                 })
                 .buffered(
                     self.config
-                        .n_readers_pre_init
+                        .n_readers_pre_init()
                         .min(self.config.sources.len()),
                 )
         };
@@ -218,9 +216,7 @@ impl MultiScanTaskInitializer {
         let projected_file_schema = self.config.projected_file_schema.clone();
         let full_file_schema = self.config.full_file_schema.clone();
         let num_pipelines = self.config.num_pipelines();
-        let max_concurrent_scans = num_pipelines
-            .min(sources.len())
-            .min(max_concurrent_scans_config());
+        let max_concurrent_scans = self.config.max_concurrent_scans();
 
         let (started_reader_tx, started_reader_rx) =
             tokio::sync::mpsc::channel(max_concurrent_scans.max(2) - 1);
@@ -312,13 +308,6 @@ impl ReaderStarter {
             // Set to IdxSize::MAX if we expect it to be unused. This way if it is incorrectly being
             // used it should cause an error.
             current_row_position = IdxSize::MAX;
-        }
-
-        if verbose {
-            eprintln!(
-                "[ReaderStarter]: max_concurrent_scans: {}",
-                max_concurrent_scans
-            )
         }
 
         let wait_group = WaitGroup::default();