* Fixes https://github.com/pola-rs/polars/issues/22683

There have been reports of performance regressions from users that scan 10,000's of small parquet files from the cloud. This is likely due to some default settings in the new-streaming I/O scans. In this PR we address this by:

* Increasing the default reader pre-initialization (i.e. parquet metadata fetch/decode) limit from 3 to `num_pipelines + 3`
  * Adds a `POLARS_NUM_READERS_PRE_INIT` environment variable to allow this to be controlled
  * Adds a verbose print line so that we can see the value used in the logs
* Changing the default concurrent reader limit from 8 to `num_pipelines`

### Benchmarks

* System: AWS EC2 `c6gn.8xlarge` (32 cores)
* Query: Full scan 1,000 parquet files (10,000 rows / ~1.1MB each)
* Build profile: Debug

| Configuration | # readers pre-init | # concurrent readers | Approximate throughput (MiB/s) |
|------|--------------------|----------------------|--------------------------------|
| Existing default for 1.29.0 | 3                  | 8                    | 60.0                           |
| Default for this branch | 35                 | 32                   | 130.0                          |